# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API",
    "version": "1.0.0",
    "description": "NodeJS/Express API"
  },
  "components": {
    "securitySchemes": {
      "OAuth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://ouath2.server.com",
            "scopes": {}
          }
        }
      }
    },
    "schemas": {
      "genericId": {
        "type": "string",
        "pattern": "d{10}",
        "maxLength": 10
      },
      "HealthcheckResponseBody": {
        "type": "object",
        "properties": {
          "buildDate": {
            "type": "string"
          },
          "buildSha": {
            "type": "string"
          },
          "buildNumber": {
            "type": "string"
          }
        },
        "required": [
          "buildDate",
          "buildSha",
          "buildNumber"
        ],
        "additionalProperties": false
      },
      "ApiResponse": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "data": {
            "additionalProperties": false,
            "type": "object"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiError"
            },
            "maxItems": 1
          },
          "meta": {
            "additionalProperties": false,
            "type": "object"
          }
        }
      },
      "ApiError": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "message": {
            "pattern": "^.*$",
            "type": "string",
            "maxLength": 1000
          },
          "detail": {
            "additionalProperties": false,
            "type": "object"
          },
          "parent": {
            "additionalProperties": false,
            "type": "object"
          }
        },
        "required": [
          "message"
        ]
      }
    },
    "responses": {
      "errorResponse": {
        "description": "Execution Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "defaultResponse": {
        "description": "Endpoint",
        "content": {
          "text/plain": {
            "schema": {
              "type": "string",
              "pattern": "^.*$",
              "maxLength": 9999999
            }
          }
        }
      }
    },
    "headers": {
      "XRequestID": {
        "description": "Request correlation identifier",
        "schema": {
          "type": "string"
        },
        "required": true
      }
    },
    "parameters": {
      "XRequestID": {
        "name": "X-Request-ID",
        "in": "header",
        "description": "Request correlation identifier",
        "schema": {
          "type": "string"
        },
        "required": true
      }
    }
  },
  "security": [
    {
      "OAuth": []
    }
  ],
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Server"
    }
  ],
  "paths": {
    "/v1/healthcheck/": {
      "get": {
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Successful healthcheck response",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthcheckResponseBody"
                }
              }
            }
          }
        }
      }
    },
    "/v2/healthcheck/": {
      "get": {
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Successful healthcheck response",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthcheckResponseBody"
                }
              }
            }
          }
        }
      }
    },
    "/v3/healthcheck/": {
      "get": {
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Successful healthcheck response",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthcheckResponseBody"
                }
              }
            }
          }
        }
      }
    },
    "/v3.1-beta1/healthcheck/": {
      "get": {
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Successful healthcheck response",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthcheckResponseBody"
                }
              }
            }
          }
        }
      }
    },
    "/v3.1/healthcheck/": {
      "get": {
        "tags": [
          "api"
        ],
        "responses": {
          "200": {
            "description": "Successful healthcheck response",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthcheckResponseBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/test1": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest1",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v2/test1": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest1",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v3/test1": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest1",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v3.1-beta1/test1": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest1",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v3.1/test1": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest1",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v1/test1/second": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest1",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v2/test1/second": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest1",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v3/test1/second": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest1",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v3.1-beta1/test1/second": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest1",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v3.1/test1/second": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest1",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v1/test1/third/{value}": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest1",
        "parameters": [
          {
            "name": "value",
            "in": "path",
            "description": "Value to return",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v2/test1/third/{value}": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest1",
        "parameters": [
          {
            "name": "value",
            "in": "path",
            "description": "Value to return",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v3.1-beta1/test1/fourth/{value}": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest1",
        "parameters": [
          {
            "name": "value",
            "in": "path",
            "description": "Value to return",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v3.1/test1/fourth/{value}": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest1",
        "parameters": [
          {
            "name": "value",
            "in": "path",
            "description": "Value to return",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v1/test2": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v2/test2": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v3/test2": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v3.1-beta1/test2": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v3.1/test2": {
      "get": {
        "summary": "Gets a test value",
        "operationId": "GetTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v1/test2/doPut": {
      "put": {
        "summary": "Gets a test value",
        "operationId": "PutTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v2/test2/doPut": {
      "put": {
        "summary": "Gets a test value",
        "operationId": "PutTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v3/test2/doPut": {
      "put": {
        "summary": "Gets a test value",
        "operationId": "PutTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v3.1-beta1/test2/doPut": {
      "put": {
        "summary": "Gets a test value",
        "operationId": "PutTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v3.1/test2/doPut": {
      "put": {
        "summary": "Gets a test value",
        "operationId": "PutTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v1/test2/doPost": {
      "post": {
        "summary": "Gets a test value",
        "operationId": "PostTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v2/test2/doPost": {
      "post": {
        "summary": "Gets a test value",
        "operationId": "PostTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v3/test2/doPost": {
      "post": {
        "summary": "Gets a test value",
        "operationId": "PostTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v3.1-beta1/test2/doPost": {
      "post": {
        "summary": "Gets a test value",
        "operationId": "PostTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v3.1/test2/doPost": {
      "post": {
        "summary": "Gets a test value",
        "operationId": "PostTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v1/test2/doDelete/{value}": {
      "delete": {
        "summary": "Gets a test value",
        "operationId": "DeleteTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        }
      }
    },
    "/v2/test2/doDelete/{value}": {
      "delete": {
        "summary": "Gets a test value",
        "operationId": "DeleteTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v3/test2/doDelete/{value}": {
      "delete": {
        "summary": "Gets a test value",
        "operationId": "DeleteTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v3.1-beta1/test2/doDelete/{value}": {
      "delete": {
        "summary": "Gets a test value",
        "operationId": "DeleteTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    },
    "/v3.1/test2/doDelete/{value}": {
      "delete": {
        "summary": "Gets a test value",
        "operationId": "DeleteTest2",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/errorResponse"
          },
          "default": {
            "$ref": "#/components/responses/defaultResponse"
          }
        },
        "deprecated": true
      }
    }
  },
  "tags": []
}